chore: clean up a couple of TypeScript warnings
