+++ packages/superset-ui-chart-controls/src/index.ts
@@ -11,7 +11,7 @@ export const internalSharedControls = sharedControlsModule;
 export const sections = sectionModules;
 export { D3_FORMAT_DOCS, D3_FORMAT_OPTIONS, D3_TIME_FORMAT_OPTIONS } from './utils/D3Formatting';
 export { formatSelectOptions, formatSelectOptionsForRange } from './utils/selectOptions';
-export * from './utils/mainMetric';
+export { default as mainMetric } from './utils/mainMetric';
 export * from './utils/expandControlConfig';
 
 export * from './components/InfoTooltipWithTrigger';
